# Conflicts:
#	evolution-common-build/pom.xml
#	evolution/pom.xml
#	evolution/src/main/java/com/evolving/nglm/core/EvolutionSetup.java
#	evolution/src/main/java/com/evolving/nglm/evolution/ThirdPartyManager.java
#	evolution/src/main/java/com/evolving/nglm/evolution/reports/contactpolicyconfig/ContactPolicyConfigurationReportDriver.java
#	evolution/src/main/java/com/evolving/nglm/evolution/reports/journeyimpact/JourneyImpactReportDriver.java
#	evolution/src/main/java/com/evolving/nglm/evolution/reports/journeys/JourneysReportDriver.java
#	evolution/src/main/java/com/evolving/nglm/evolution/reports/product/ProductReportDriver.java
#	evolution/src/main/resources/bin/evolution-prepare-stack.sh
#	evolution/src/main/resources/config/evolution-prometheus-application.yml
#	evolution/src/main/resources/config/grafana-gui-monitoring-dashboard.json
#	evolution/src/main/resources/support/README-evolution
#	pom.xml

+++ evolution/src/main/java/com/evolving/nglm/evolution/MetricHistory.java
@@ -481,7 +481,7 @@ public class MetricHistory
     byte[] packedMonthlyBuckets = valueStruct.getBytes("monthlyBuckets");
     long allTimeBucket = valueStruct.getInt64("allTimeBucket");
     MetricHistoryMode metricHistoryMode = (schemaVersion >= 2) ? MetricHistoryMode.fromInternalRepresentation(valueStruct.getInt32("metricHistoryMode")) : MetricHistoryMode.Standard;
-    int tenantID = schema.field("tenantID") != null ? valueStruct.getInt32("tenantID") : 1;
+    int tenantID = schema.field("tenantID") != null ? valueStruct.getInt16("tenantID") : 1;
     
     //
     //  unpack buckets
